+++ deformo/Loader.h
@@ -4,7 +4,7 @@
 #include <string>
 
 namespace loader {
-void ReadTetgenVertexFile(Eigen::MatrixXd& V, const std::string& node_file);
-void ReadTetgenFaceFile(Eigen::MatrixXd& F, const std::string& face_file);
-void ReadTetgenEleFile(Eigen::MatrixXd& T, const std::string& ele_file);
+void ReadTetgenVertexFile(Eigen::MatrixXf& V, const std::string& node_file);
+void ReadTetgenFaceFile(Eigen::MatrixXf& F, const std::string& face_file);
+void ReadTetgenEleFile(Eigen::MatrixXf& T, const std::string& ele_file);
 }  // namespace loader
